+++ changelog.md
@@ -4,7 +4,7 @@
 
 ### 0.1.0 `17 février 2024`
 
--   Copie du starter Vite/React pour réinitialiser le projet
+-   Copy of starter Vite/React to reset the project
 
 ### 0.2.0 `17 février 2024`
 
@@ -17,5 +17,13 @@
 -   Delete all starter content
 -   Change Font-Family of project from Poppins to Inknut Antiqua
 -   Add alias to the project
+
+### 0.2.1 `17 février 2024`
+
+-   Add react-router-dom to the project
+-   Add "specials" to the alias
+-   Create Component "AppRouter" in special, wich contain React Router in futur
+-   Transform export default to named export
+-   Add slice and add it to redux
 
 </details>